move the two functions into datadir.py
