+++ scripts/housekeeping/datadir.py
@@ -2,6 +2,7 @@ import os
 import platform
 
 from scripts.housekeeping.version import get_version_info
+from scripts.utility import logger
 
 
 def setup_data_dir():
@@ -54,3 +55,24 @@ def get_temp_dir():
 
 def get_saved_images_dir():
     return get_data_dir() + "/saved_images"
+
+
+def open_data_dir():
+    if platform.system() == "Darwin":
+        subprocess.Popen(["open", "-R", get_data_dir()])
+    elif platform.system() == "Windows":
+        os.startfile(get_data_dir())  # pylint: disable=no-member
+    elif platform.system() == "Linux":
+        try:
+            subprocess.Popen(["xdg-open", get_data_dir()])
+        except OSError:
+            logger.exception("Failed to call to xdg-open.")
+
+
+def open_url(url: str):
+    if platform.system() == "Darwin":
+        subprocess.Popen(["open", "-u", url])
+    elif platform.system() == "Windows":
+        os.system(f"start \"\" {url}")
+    elif platform.system() == "Linux":
+        subprocess.Popen(["xdg-open", url])
