+++ scripts/screens/SettingsScreen.py
@@ -17,7 +17,8 @@ from scripts.game_structure.ui_elements import (
     UISurfaceImageButton,
     UIImageHorizontalSlider,
 )
-from scripts.utility import get_text_box_theme, ui_scale, ui_scale_dimensions, open_data_dir
+from scripts.housekeeping.datadir import open_data_dir
+from scripts.utility import get_text_box_theme, ui_scale, ui_scale_dimensions
 from .Screens import Screens
 from ..game_structure.audio import music_manager, sound_manager
 from ..game_structure.screen_settings import (
